+++ packages/org/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.2.50] - 2020-01-31
 ## [5.2.49] - 2020-01-17
 ### Changed
  - renamed caption of sextype from 'Sex' to 'Gender'